feat: Add Footer component
This commit adds a new Footer component to the project. The Footer component is responsible for displaying the copyright information at the bottom of the page. It includes a simple text element with the copyright symbol and the year. This component will be used across multiple pages in the application.

The Footer component is implemented as a functional component using React. It is styled using Tailwind CSS classes to achieve the desired appearance.

This commit also includes the necessary import statements and exports the Footer component as the default export.